Inlined package lists in apt tasks and removed corresponding default variables
The list of packages to be installed during role execution is not meant to be
given or changed by the user. Storing these package lists as role
(default) variables is misleading and not idiomatic. Hence the lists have
been inlined in tasks/setup-*.yml and the corresponding variables have been
removed from defaults/main.yml.